refactor(app): one last up import
