+++ app/src/local-resources/labware/utils/index.ts
@@ -1,1 +1,2 @@
 export * from './getAllDefinitions'
+export * from './labwareImages'
