+++ app/src/local-resources/labware/utils/labwareImages.ts
